report_builder/requirements.txt
numpy==1.25.0
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tzdata==2023.3
